requirements.txt
Django==4.2.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
psycopg2~=2.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt
django-filter